mint/archive/requirements_new.txt
cvxpy==1.4.1
fpdf2==2.7.6
imageio==2.32.0
joblib==1.3.2
JPype1==1.4.1
JPype1==1.4.1
matplotlib==3.4.3
matplotlib==3.8.0
nd2==0.8.1
numpy==1.26.2
numpy==1.26.2
pandas==1.5.2
PyYAML==6.0
PyYAML==6.0.1
scikit_posthocs==0.8.0
scipy==1.11.4
scs==3.2.4.post1
seaborn==0.13.0
setuptools==64.0.2
setuptools==65.6.3
trackpy==0.6.1
